+++ src/include.hpp
@@ -75,4 +75,5 @@
 #include "driver/sys/DS1307.hpp"
 #include "driver/sys/SDCard.hpp"
 #include "driver/disp/ST7701.hpp"
-#include "driver/comm/PN532.hpp"
+#include "driver/comm/nfc/PN532.hpp"
+#include "driver/comm/eth/W5500.hpp"